+++ primitives/header-chain/Cargo.toml
@@ -9,12 +9,10 @@ license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 [dependencies]
 codec = { package = "parity-scale-codec", version = "1.3.1", default-features = false }
 serde = { version = "1.0", optional = true }
-finality-grandpa = { version = "0.12.3", default-features = false }
 
 # Substrate Dependencies
 
 sp-finality-grandpa = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-sp-keyring = { git = "https://github.com/paritytech/substrate.git", branch = "master", default-features = false }
 sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
